map icon added - menu page added
