+++ beer_me/Resources/layout/BreweryList.cs
@@ -1,0 +1,26 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+using Android.App;
+using Android.Content;
+using Android.OS;
+using Android.Runtime;
+using Android.Views;
+using Android.Widget;
+
+namespace beer_me
+{
+	[Activity(Label = "BreweryList")]
+	public class BreweryList : Activity
+	{
+		protected override void OnCreate(Bundle savedInstanceState)
+		{
+			base.OnCreate(savedInstanceState);
+
+			// Create your application here
+		}
+	}
+}
