+++ beer_me/Resources/layout/Main.axml
@@ -2,13 +2,23 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:minWidth="25px"
-    android:minHeight="25px">
-    <ListView
-        android:id="@+id/breweryListView"
-        android:minWidth="25px"
-        android:minHeight="25px"
+    android:layout_height="match_parent">
+    <ImageView
+        android:src="@drawable/asba_logo"
+        android:layout_width="350dp"
+        android:layout_height="350dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginBottom="30dp"
+        android:layout_marginTop="30dp"
+        android:id="@+id/asbaLogo" />
+    <Button
+        android:text="Find me a Brew"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="wrap_content"
+        android:id="@+id/findBrewButton" />
+    <Button
+        android:text="Brewery List"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:id="@+id/breweryListButton" />
 </LinearLayout>
