+++ beer_me/Resources/layout/SingleBrewery.axml
@@ -2,41 +2,33 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:actionBarNavMode="tabs">
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent">
     <ImageView
-        android:layout_width="300dp"
-        android:layout_height="300dp"
+        android:layout_width="200dp"
+        android:layout_height="200dp"
         android:layout_gravity="center_horizontal"
         android:id="@+id/breweryImage" />
-    <Button
-        android:text="View Map"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
-        android:layout_marginBottom="20dp"
+    <ImageView
+        android:src="@drawable/map_icon"
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginBottom="30dp"
         android:id="@+id/mapView" />
-    <TextView
-        android:textAppearance="?android:attr/textAppearanceLarge"
+    <ScrollView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
         android:layout_marginBottom="20dp"
-        android:id="@+id/breweryName" />
-    <Button
-        android:text="GPS"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/gps" />
-    <TextView
-        android:textAppearance="?android:attr/textAppearanceSmall"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:textSize="20sp"
-        android:singleLine="false"
-        android:id="@+id/breweryDescription" />
+        android:id="@+id/scrollView1">
+        <TextView
+            android:textAppearance="?android:attr/textAppearanceSmall"
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="20dp"
+            android:layout_marginRight="20dp"
+            android:textSize="20sp"
+            android:singleLine="false"
+            android:id="@+id/breweryDescription" />
+    </ScrollView>
 </LinearLayout>
