+++ beer_me/beer_me.csproj
@@ -90,6 +90,7 @@
     <Compile Include="BreweryDataService.cs" />
     <Compile Include="SingleBrewery.cs" />
     <Compile Include="BeerMe.cs" />
+    <Compile Include="BreweryList.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\AboutResources.txt" />
@@ -115,7 +116,6 @@
     <AndroidResource Include="Resources\drawable\logo_two_sergents.png" />
     <AndroidResource Include="Resources\drawable\logo_troubled_monk.png" />
     <AndroidResource Include="Resources\drawable\logo_tool_shed.png" />
-    <AndroidResource Include="Resources\drawable\logo_theortically.jpg" />
     <AndroidResource Include="Resources\drawable\logo_something_brewing.png" />
     <AndroidResource Include="Resources\drawable\logo_ribstone.png" />
     <AndroidResource Include="Resources\drawable\logo_olds_college.png" />
@@ -126,7 +126,6 @@
     <AndroidResource Include="Resources\drawable\logo_half_hitch.png" />
     <AndroidResource Include="Resources\drawable\logo_grizzly_paw.png" />
     <AndroidResource Include="Resources\drawable\logo_grain_bin.png" />
-    <AndroidResource Include="Resources\drawable\logo_gp.jpg" />
     <AndroidResource Include="Resources\drawable\logo_drummond.png" />
     <AndroidResource Include="Resources\drawable\logo_dandy.png" />
     <AndroidResource Include="Resources\drawable\logo_coulee.png" />
@@ -134,11 +133,18 @@
     <AndroidResource Include="Resources\drawable\logo_brewsters.png" />
     <AndroidResource Include="Resources\drawable\logo_blindman.png" />
     <AndroidResource Include="Resources\drawable\logo_big_rock.png" />
-    <AndroidResource Include="Resources\drawable\logo_bent_stick.jpg" />
     <AndroidResource Include="Resources\drawable\logo_bench_creek.png" />
     <AndroidResource Include="Resources\drawable\logo_banff_ave.png" />
     <AndroidResource Include="Resources\drawable\logo_alley_cat.png" />
     <AndroidResource Include="Resources\layout\BeerMe.axml" />
+    <AndroidResource Include="Resources\layout\BreweryList.axml" />
+    <AndroidResource Include="Resources\drawable\map_icon.png" />
+    <AndroidResource Include="Resources\drawable\instagram_logo.png" />
+    <AndroidResource Include="Resources\drawable\facebook_logo.png" />
+    <AndroidResource Include="Resources\drawable\logo_theortically.jpg" />
+    <AndroidResource Include="Resources\drawable\logo_gp.jpg" />
+    <AndroidResource Include="Resources\drawable\logo_bent_stick.jpg" />
+    <AndroidResource Include="Resources\drawable\asba_logo.png" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Resources\drawable\" />
